feat: making tom less mad & fixing bugs with render service and mesh service

+++ src/engine/graphics/Mesh.cpp
@@ -1,11 +1,11 @@
 #include "Mesh.h"
 
 Mesh::Mesh() {
-  this->verts = std::vector<double>();
+  this->verts = std::vector<float>();
   this->faces = std::vector<int>();
 }
 
-Mesh::Mesh(std::vector<double> verts, std::vector<int> faces) {
+Mesh::Mesh(std::vector<float> verts, std::vector<int> faces) {
   this->verts = verts;
   this->faces = faces;
 }
